Add files via upload
parse_manifest function takes the file id of an open file and returns the head of a linked list of a struct that contains the info of every file in the manifest, could be used in a lot of places i think. the manifest itself needs a newline between every piece of information. i uploaded an example